+++ objectstore/BackendRados.cpp
@@ -27,6 +27,7 @@
 #include "BackendRados.hpp"
 #include "common/exception/Errnum.hpp"
 #include "common/exception/NoSuchObject.hpp"
+#include "common/exception/TimeoutException.hpp"
 #include "common/log/LogContext.hpp"
 #include "common/threading/MutexLocker.hpp"
 #include "common/Timer.hpp"