+++ scheduler/Scheduler.hpp
@@ -333,10 +333,10 @@ public:
    * @param logicalLibraryName library for the drive we are scheduling
    * @param driveName name of the drive we are scheduling
    * @param lc log context
-   * @param globalLockTimeout_us global lock timeout
+   * @param timeout_us get next mount timeout
    * @return unique pointer to the tape mount structure. Next step for the user will be find which type of mount this is.
    */
-  std::unique_ptr<TapeMount> getNextMount(const std::string &logicalLibraryName, const std::string &driveName, log::LogContext & lc, uint64_t globalLockTimeout_us = 0);
+  std::unique_ptr<TapeMount> getNextMount(const std::string &logicalLibraryName, const std::string &driveName, log::LogContext & lc, uint64_t timeout_us = 0);
 
   /**
    * A function returning